EduWeb/pyproject.toml
[tool.poetry]
name = "eduweb"
version = "0.1.0"
description = ""
authors = ["eseminara <eze.seminara@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
dj-auth = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
